Refactor RandomAgent class to inherit from base class
