+++ src/main/java/controller/RandomAgent.java
@@ -24,23 +24,10 @@ package controller;
 
 import core.game.Observation;
 import core.game.StateObservation;
-import core.player.AbstractPlayer;
-import core.vgdl.VGDLRegistry;
-import ontology.Types;
-import org.yaml.snakeyaml.Yaml;
-import org.yaml.snakeyaml.constructor.Constructor;
 import tools.ElapsedCpuTimer;
 import tools.Vector2d;
 
-import java.io.*;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.*;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 /**
  * Random agent class. It represents an agent that has a random behaviour. This
@@ -48,24 +35,8 @@ import java.util.stream.Stream;
  *
  * @author Vladislav Nikolov Vasilev
  */
-public class RandomAgent extends AbstractPlayer {
-    // The following attributes can be modified
-    protected static String gameConfigFile;
-
-    // Game information data structure (loaded from a .yaml file) and file path
-    protected GameInformation gameInformation;
-
-    // Set of connections between cells
+public class RandomAgent extends AbstractRandomAgent {
     protected Set<String> connectionSet;
-    protected Map<String, Set<String>> gameElementVars;
-
-    protected Random randomGenerator;
-    protected Map<Types.ACTIONS, String> actionCorrespondence;
-
-    protected static List<String> executedActions = new ArrayList<>();
-    protected static List<String> gamePredicates = new ArrayList<>();
-
-    protected boolean isGameOverDetected;
 
     /**
      * Class constructor. Creates a new random agent.
@@ -74,104 +45,8 @@ public class RandomAgent extends AbstractPlayer {
      * @param elapsedCpuTimer  Elapsed CPU time.
      */
     public RandomAgent(StateObservation stateObservation, ElapsedCpuTimer elapsedCpuTimer) {
-      // Load game information
-      Yaml yaml = new Yaml(new Constructor(GameInformation.class));
-
-      try {
-          InputStream inputStream = new FileInputStream(new File(RandomAgent.gameConfigFile));
-          this.gameInformation = yaml.load(inputStream);
-      } catch (FileNotFoundException e) {
-          System.out.println(e.getStackTrace());
-      }
-
-      // Initialize PDDL related variables
-      this.gameElementVars = this.extractVariablesFromPredicates();
+      super(stateObservation, elapsedCpuTimer);
       this.connectionSet = this.generateConnectionPredicates(stateObservation);
-
-      this.randomGenerator = new Random();
-
-      this.actionCorrespondence = new HashMap<>();
-
-      this.actionCorrespondence.put(Types.ACTIONS.ACTION_UP, "MOVE_UP");
-      this.actionCorrespondence.put(Types.ACTIONS.ACTION_DOWN, "MOVE_DOWN");
-      this.actionCorrespondence.put(Types.ACTIONS.ACTION_LEFT, "MOVE_LEFT");
-      this.actionCorrespondence.put(Types.ACTIONS.ACTION_RIGHT, "MOVE_RIGHT");
-      this.actionCorrespondence.put(Types.ACTIONS.ACTION_USE, "USE");
-
-      this.isGameOverDetected = false;
-    }
-
-    /**
-     * Method called in each turn that returns the next action that the agent
-     * will execute. It is responsible for controlling the agent's behaviour.
-     *
-     * @param stateObservation State observation of the game.
-     * @param elapsedCpuTimer  Elapsed CPU time
-     * @return Returns the action that will be executed by the agent in the
-     * current turn.
-     */
-    @Override
-    public Types.ACTIONS act(StateObservation stateObservation, ElapsedCpuTimer elapsedCpuTimer) {
-      // Select a random action between the available ones
-      ArrayList<Types.ACTIONS> availableActions = stateObservation.getAvailableActions();
-
-      int index = randomGenerator.nextInt(availableActions.size());
-      Types.ACTIONS action = availableActions.get(index);
-
-      // Translate game state to PDDL predicates
-      String predicates = this.translateGameStateToPDDL(stateObservation);
-
-      if (this.isGameOverDetected) {
-        RandomAgent.gamePredicates.remove(RandomAgent.gamePredicates.size() - 1);
-        this.isGameOverDetected = false;
-      }
-
-      RandomAgent.gamePredicates.add(predicates);
-
-      // Instantiate action
-      String instantiatedAction;
-      String actionStr = this.actionCorrespondence.get(action);
-
-      // Use forward model to extract information about the current action
-      // and the next game state
-      StateObservation nextState = stateObservation.copy();
-      nextState.advance(action);
-
-      int numResources = this.getNumberResources(stateObservation);
-      //System.out.println(stateObservation.getResourcesPositions(stateObservation.getAvatarPosition())[0]);
-
-      //System.out.println(numResources);
-
-      if (!stateObservation.getAvatarOrientation().equals(nextState.getAvatarOrientation())) {
-        actionStr = actionStr.replace("MOVE", "TURN");
-        instantiatedAction = this.createAvatarAction(actionStr);
-      } else if (action.equals(Types.ACTIONS.ACTION_USE)) {
-        instantiatedAction = this.createAvatarAction(actionStr);
-      } else {
-        Vector2d currentAvatarPos = stateObservation.getAvatarPosition();
-        
-        int x = (int)currentAvatarPos.x / stateObservation.getBlockSize();
-        int y = (int)currentAvatarPos.y / stateObservation.getBlockSize();
-
-        boolean isResourcePicked = numResources != this.getNumberResources(nextState);
-
-        instantiatedAction = this.createMoveAction(stateObservation, actionStr, x, y, isResourcePicked);
-      }
-
-      for (int i = 0; i < 10 && !this.isGameOverDetected; i++) {
-        if (nextState.isGameOver()) {
-          String nextTurnPredicates = this.translateGameStateToPDDL(nextState);
-          RandomAgent.gamePredicates.add(nextTurnPredicates);
-          this.isGameOverDetected = true;
-        }
-
-        nextState = stateObservation.copy();
-        nextState.advance(action);
-      }
-
-      RandomAgent.executedActions.add(instantiatedAction);
-
-      return action;
     }
 
     /**
@@ -181,7 +56,7 @@ public class RandomAgent extends AbstractPlayer {
      * @return Returns an uppercase String that contains the PDDL predicates.
      * The type of each object is specified in the predicate.
      */
-    public String translateGameStateToPDDL(StateObservation stateObservation) {
+    protected String translateGameStateToPDDL(StateObservation stateObservation) {
       List<String> predicates = new ArrayList<>();
 
       // Get the observations of the game state as elements of the VGDDLRegistry
@@ -261,62 +136,94 @@ public class RandomAgent extends AbstractPlayer {
     }
 
     /**
-     * Method that translates a game state observation to a matrix of strings which
-     * represent the elements of the game in each position according to the VGDDL
-     * registry. There can be more than one game element in each position.
+     * Method used to create an action instance of a movement action. An instance
+     * of a movement action follows this pattern:
+     * (MOV_ACTION AVATAR_VARIABLE - AVATAR_TYPE CURRENT_CELL - CELL_TYPE NEXT_CELL - CELL_TYPE).
      *
-     * @param stateObservation State observation of the game.
-     * @return Returns a matrix containing the elements of the game in each position.
-     */
-    public HashSet<String>[][] getGameElementsMatrix(StateObservation stateObservation) {
-        // Get the current game state
-        ArrayList<Observation>[][] gameState = stateObservation.getObservationGrid();
+     * @param stateObs Current state observation.
+     * @param actionStr String that contains the movement action that is going
+     * to be instantiated.
+     * @param currentX Position of the avatar on the X-axis.
+     * @param currentY Position of the avatar on the Y-axis.
+     * @param isResourcePicked Boolean that tells if a resource will be picked
+     * when the action is executed.
+     * @return Returns a String containing the instantiated action.
+     */ 
+    protected String createMoveAction(StateObservation stateObs, String actionStr,
+        int currentX, int currentY, boolean isResourcePicked) {
+      String cellVariable = this.gameInformation.cellVariable;
+      String cellType = this.gameInformation.variablesTypes.get(cellVariable);
 
-        // Get the number of X tiles and Y tiles
-        final int X_MAX = gameState.length, Y_MAX = gameState[0].length;
+      String avatarVariable = this.gameInformation.avatarVariable;
+      String avatarType = this.gameInformation.variablesTypes.get(avatarVariable);
 
-        // Create a new matrix, representing the game's map
-        HashSet<String>[][] gameStringMap = new HashSet[X_MAX][Y_MAX];
+      int nextX = currentX;
+      int nextY = currentY;
+      
+      switch(actionStr) {
+        case "MOVE_UP":
+          nextY--;
+          break;
+        case "MOVE_DOWN":
+          nextY++;
+          break;
+        case "MOVE_RIGHT":
+          nextX++;
+          break;
+        case "MOVE_LEFT":
+          nextX--;
+          break;
+      }
 
-        // Iterate over the map and transform the observations in a [x, y] cell
-        // to a HashSet of Strings. In case there's no observation, add a
-        // "background" string. The VGDLRegistry contains the needed information
-        // to transform the StateObservation to a matrix of sets of Strings.
-        for (int y = 0; y < Y_MAX; y++) {
-            for (int x = 0; x < X_MAX; x++) {
-                gameStringMap[x][y] = new HashSet<>();
+      // Instantiate current cell and next cell objects
+      String currentCell = String.format("%s_%d_%d", cellVariable, currentX, currentY);
+      String nextCell = String.format("%s_%d_%d", cellVariable, nextX, nextY);
+      String instantiatedAction;
 
-                if (gameState[x][y].size() > 0) {
-                    for (int i = 0; i < gameState[x][y].size(); i++) {
-                        gameStringMap[x][y].add(this.getGameElementFromObservation(gameState[x][y].get(i)));
-                    }
-                } else {
-                    gameStringMap[x][y].add("background");
-                }
+      if (isResourcePicked) {
+        actionStr += "_PICK_RESOURCE";
+
+        int blockSize = stateObs.getBlockSize();
+
+        Vector2d resourcePosition = new Vector2d(nextX * blockSize, nextY * blockSize);
+        ArrayList<Observation>[] resources = stateObs.getResourcesPositions(resourcePosition);
+
+        Observation resourceObservation = null;
+
+        for (int i = 0; i < resources.length; i++) {
+          for (Observation obs: resources[i]) {
+            if (obs.sqDist == 0.0) {
+              resourceObservation = obs;
             }
+          }
         }
 
-        return gameStringMap;
-    }
+        String gameElement = this.getGameElementFromObservation(resourceObservation);
+        String resourceObject = "";
 
-    public static void setGameConfigFile(String path) {
-        RandomAgent.gameConfigFile = path;
-    }
+        for (String object: this.gameElementVars.get(gameElement)) {
+          if (!object.equals(this.gameInformation.cellVariable)) {
+            resourceObject = object;
+          }
+        }
 
-    public static void displayPostGameInformation() {
-      System.out.println("##Tasks##");
-      for (int i = 0; i < RandomAgent.executedActions.size(); i++) {
-        System.out.println(String.format("[%d, %d]: %s", i, i+1, RandomAgent.executedActions.get(i)));
+        String resourceType = this.gameInformation.variablesTypes.get(resourceObject);
+        resourceObject = String.format("%s_%d_%d", resourceObject, nextX, nextY);
+
+        instantiatedAction = String.format("(%s %s - %s %s - %s %s - %s %s - %s)",
+            actionStr, avatarVariable, avatarType, currentCell, cellType, nextCell,
+            cellType, resourceObject, resourceType)
+          .replace("?", "")
+          .toUpperCase();
+      } else {
+        instantiatedAction = String.format("(%s %s - %s %s - %s %s - %s)",
+            actionStr, avatarVariable, avatarType, currentCell, cellType, nextCell,
+            cellType)
+          .replace("?", "")
+          .toUpperCase();
       }
 
-      System.out.println("\n\n##States##");
-      for (int i = 0; i < RandomAgent.gamePredicates.size(); i++) {
-        System.out.println(String.format("[%d]: %s\n", i, RandomAgent.gamePredicates.get(i)));
-      }
-    }
-
-    protected String getGameElementFromObservation(Observation obs) {
-      return VGDLRegistry.GetInstance().getRegisteredSpriteKey(obs.itype);
+      return instantiatedAction;
     }
 
     /**
@@ -401,161 +308,5 @@ public class RandomAgent extends AbstractPlayer {
         }
 
         return connections;
-    }
-
-    /**
-     * Method used to extract the variables from the predicates associated to a game
-     * element and associate them to game elements directly.
-     *
-     * @return Returns a mapping between the game elements and the variables
-     * associated to them.
-     */
-    private Map<String, Set<String>> extractVariablesFromPredicates() {
-      Map<String, Set<String>> varsFromPredicates = new HashMap<>();
-
-      // Pattern that matches a variable
-      Pattern variablePattern = Pattern.compile("\\?[a-zA-Z]+");
-
-      // Iterate over all the pairs <game element: [predicates]>
-      for (Map.Entry<String, ArrayList<String>> entry : this.gameInformation.gameElementsCorrespondence.entrySet()) {
-          String gameObservation = entry.getKey();
-          Set<String> variables = new HashSet<>();
-
-          // Iterate over the predicates searching for variables
-          for (String observation : entry.getValue()) {
-              Matcher variableMatcher = variablePattern.matcher(observation);
-
-              while (variableMatcher.find()) {
-                  for (int i = 0; i <= variableMatcher.groupCount(); i++) {
-                      variables.add(variableMatcher.group(i));
-                  }
-              }
-          }
-
-          // Add the predicates associated to the game element
-          varsFromPredicates.put(gameObservation, variables);
-      }
-
-      return varsFromPredicates;
-    }
-
-    /**
-     * Method used to create an action instance which only involves the avatar such
-     * as an action that changes the avatar's orientation or an USE action.
-     *
-     * @param actionStr String that contains the action to be instantiated.
-     * @return Returns a String containing the instantiated action.
-     */
-    private String createAvatarAction(String actionStr) {
-      String avatarVariable = this.gameInformation.avatarVariable;
-
-      return String.format("(%s %s - %s)", actionStr, avatarVariable,
-            this.gameInformation.variablesTypes.get(avatarVariable)
-          )
-        .replace("?", "")
-        .toUpperCase();
-    }
-
-    /**
-     * Method used to create an action instance of a movement action. An instance
-     * of a movement action follows this pattern:
-     * (MOV_ACTION AVATAR_VARIABLE - AVATAR_TYPE CURRENT_CELL - CELL_TYPE NEXT_CELL - CELL_TYPE).
-     *
-     * @param stateObs Current state observation.
-     * @param actionStr String that contains the movement action that is going
-     * to be instantiated.
-     * @param currentX Position of the avatar on the X-axis.
-     * @param currentY Position of the avatar on the Y-axis.
-     * @param isResourcePicked Boolean that tells if a resource will be picked
-     * when the action is executed.
-     * @return Returns a String containing the instantiated action.
-     */ 
-    private String createMoveAction(StateObservation stateObs, String actionStr,
-        int currentX, int currentY, boolean isResourcePicked) {
-      String cellVariable = this.gameInformation.cellVariable;
-      String cellType = this.gameInformation.variablesTypes.get(cellVariable);
-
-      String avatarVariable = this.gameInformation.avatarVariable;
-      String avatarType = this.gameInformation.variablesTypes.get(avatarVariable);
-
-      int nextX = currentX;
-      int nextY = currentY;
-      
-      switch(actionStr) {
-        case "MOVE_UP":
-          nextY--;
-          break;
-        case "MOVE_DOWN":
-          nextY++;
-          break;
-        case "MOVE_RIGHT":
-          nextX++;
-          break;
-        case "MOVE_LEFT":
-          nextX--;
-          break;
-      }
-
-      // Instantiate current cell and next cell objects
-      String currentCell = String.format("%s_%d_%d", cellVariable, currentX, currentY);
-      String nextCell = String.format("%s_%d_%d", cellVariable, nextX, nextY);
-      String instantiatedAction;
-
-      if (isResourcePicked) {
-        actionStr += "_PICK_RESOURCE";
-
-        int blockSize = stateObs.getBlockSize();
-
-        Vector2d resourcePosition = new Vector2d(nextX * blockSize, nextY * blockSize);
-        ArrayList<Observation>[] resources = stateObs.getResourcesPositions(resourcePosition);
-
-        Observation resourceObservation = null;
-
-        for (int i = 0; i < resources.length; i++) {
-          for (Observation obs: resources[i]) {
-            if (obs.sqDist == 0.0) {
-              resourceObservation = obs;
-            }
-          }
-        }
-
-        String gameElement = this.getGameElementFromObservation(resourceObservation);
-        String resourceObject = "";
-
-        for (String object: this.gameElementVars.get(gameElement)) {
-          if (!object.equals(this.gameInformation.cellVariable)) {
-            resourceObject = object;
-          }
-        }
-
-        String resourceType = this.gameInformation.variablesTypes.get(resourceObject);
-        resourceObject = String.format("%s_%d_%d", resourceObject, nextX, nextY);
-
-        instantiatedAction = String.format("(%s %s - %s %s - %s %s - %s %s - %s)",
-            actionStr, avatarVariable, avatarType, currentCell, cellType, nextCell,
-            cellType, resourceObject, resourceType)
-          .replace("?", "")
-          .toUpperCase();
-      } else {
-        instantiatedAction = String.format("(%s %s - %s %s - %s %s - %s)",
-            actionStr, avatarVariable, avatarType, currentCell, cellType, nextCell,
-            cellType)
-          .replace("?", "")
-          .toUpperCase();
-      }
-
-      return instantiatedAction;
-    }
-
-    private int getNumberResources(StateObservation stateObs) {
-      int numResources = 0;
-
-      ArrayList<Observation>[] resources = stateObs.getResourcesPositions();
-
-      for (int i = 0; i < resources.length; i++) {
-        numResources += resources[i].size();
-      }
-
-      return numResources;
     }
 }
